mise_app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
google-cloud-bigquery==3.25.0
python-dateutil==2.9.0.post0
pydantic==2.9.2
python-multipart==0.0.20
requests==2.32.3
google-auth==2.34.0
word2number==1.1
rapidfuzz==3.9.7
Unidecode==1.3.8
pandas==2.2.3
openpyxl==3.1.5
anthropic==0.43.0
gspread>=6.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
qrcode>=7.4.0
pillow>=10.0.0
google-api-python-client>=2.100.0
